requirements.txt
python-dotenv>=1.0.0
websockets>=11.0.3 
mcp>=1.8.1
pydantic>=2.11.4
mcp-proxy>=0.8.2
pillow>=10.4.0

